Initial Rewrite

+++ gear_mini_v0_001/sensors.h
@@ -4,10 +4,10 @@
 #include "Arduino.h"
 
 void initialiseADC();
-void readAnalog(uint8_t AinCH);
-void readDigitalIn(uint8_t DinCH);
-void getAnalogBYoffset(uint8_t AinCH, uint8_t readloop);
+void readAnalog(byte AinCH);
+void readDigitalIn(byte DinCH);
+void getAnalogBYoffset(byte AinCH, byte readloop);
 
-unsigned int tempReading;
+uint16_t tempReading;
     
 #endif
