+++ gear_mini_v0_001/timers.h
@@ -32,9 +32,9 @@ Hence we will preload the timer with 131 cycles to leave 125 until overflow (1ms
   void oneMSInterval();
 #endif  
   
-volatile int loop100ms;
-volatile int loop250ms;
-volatile int loopSec;
+volatile int16_t loop100ms;
+volatile int16_t loop250ms;
+volatile int16_t loopSec;
 
 void initialiseTimers();
